Fix move push
